+++ addOns/reports/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Wappalyzer data to the traditional-html-plus report, if it is available.
 - Traditional JSON report
+- Automation job: outputSummary, aimed at mimicking the output of the packaged scans.
 
 ### Changed
 - Maintenance changes.